requirement.txt
atomicwrites==1.3.0
attrs==19.1.0
awscli==1.16.252
bitly-api==0.3
boto3==1.9.242
botocore==1.12.242
certifi==2019.9.11
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.5
django-cors-headers==3.1.0
django-extensions==2.2.1
django-redis==4.10.0
django-rest-swagger==2.1.2
django-storages==1.7.2
django-url-shortening==1.0.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
docutils==0.15.2
eventemitter==0.2.0
idna==2.8
importlib-metadata==0.23
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
more-itertools==7.2.0
mysqlclient==1.4.4
openapi-codec==1.3.2
packaging==19.2
Pillow==6.2.0
pluggy==0.13.0
py==1.8.0
pyasn1==0.4.7
pydotplus==2.0.2
pyee==6.0.0
PyJWT==1.7.1
pymitter==0.2.3
pyparsing==2.4.2
pyshorteners==0.6.1
pytest==5.1.3
pytest-django==3.5.1
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
PyYAML==5.1.2
redis==3.3.8
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.1
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
wcwidth==0.1.7
zipp==0.6.0
